Moved location check into login activity to fix some issues.
